Moved `AttackMoveResult` and `TurnMove` to own files
