+++ src/@types/attack-move-result.ts
@@ -1,0 +1,12 @@
+import type { BattlerIndex } from "#app/battle";
+import type { Moves } from "#enums/moves";
+import type { DamageResult } from "#app/field/pokemon";
+
+export interface AttackMoveResult {
+  move: Moves;
+  result: DamageResult;
+  damage: number;
+  critical: boolean;
+  sourceId: number;
+  sourceBattlerIndex: BattlerIndex;
+}
