+++ test/battlerTags/substitute.test.ts
@@ -1,5 +1,5 @@
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
-import type { TurnMove } from "#app/field/pokemon";
+import type { TurnMove } from "#app/@types/turn-move";
 import type { PokemonTurnData } from "#app/@types/pokemon-data";
 import type Pokemon from "#app/field/pokemon";
 import { MoveResult } from "#app/field/pokemon";
